requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
basicsr
facexlib
gfpgan
realesrgan
pytorch-lightning~=1.6.5
timm~=0.6.7
transformers
ultralytics

